feat: improve print function and normalize outputs
