+++ src/services/api.js
@@ -16,13 +16,13 @@ const api = axios.create({
 })
 
 api.interceptors.request.use(function (config) {
-  cli.action.start('Fetching..')
+  cli.action.start('Fetching')
   return config
 })
 
 api.interceptors.response.use(
   (res) => {
-    cli.action.stop('done')
+    cli.action.stop('done!')
     return res
   },
   (error) => {
