+++ src/services/flags.js
@@ -1,0 +1,13 @@
+const { flags } = require('@oclif/command')
+
+const config = require('../services/config')
+const defaultOutput = config.get('output')
+
+module.exports = {
+  output: flags.string({
+    char: 'o',
+    description: 'output type',
+    default: defaultOutput,
+    options: ['plain', 'human', 'json'],
+  }),
+}
