+++ src/services/utils.js
@@ -1,7 +1,12 @@
 const consola = require('consola')
-const { cli } = require('cli-ux')
+const table = require('text-table')
+require('console.table')
 
-function print(data, { output = 'text', format = 'table', type = 'info' }) {
+function print(data, { output } = {}) {
+  if (!data) {
+    return
+  }
+
   if (output === 'json') {
     consola.log(JSON.stringify(data, null, 2))
     return
@@ -9,30 +14,12 @@ function print(data, { output = 'text', format = 'table', type = 'info' }) {
 
   process.stdout.write('\n')
 
-  if (format === 'table') {
-    cli.table(data, {
-      id: { header: 'ID' },
-      name: {},
-      checkType: { header: 'Check Type' },
-      frequency: {},
-      locations: {
-        get: (row) => row.locations.join(','),
-      },
-      activated: {
-        get: (row) => (row.activated === true ? '✅' : '❌'),
-      },
-    })
+  if (output === 'plain') {
+    console.log(table(data.map((fields) => Object.values(fields))))
     return
   }
 
-  if (format === 'text') {
-    consola.log(
-      data.map((fields) => Object.values(fields).join(', ')).join('\n')
-    )
-    return
-  }
-
-  consola[type](data)
+  console.table(data)
 }
 
 module.exports = {
